DRY params query code
