+++ osmoutils/osmocli/query_cmd_wrap.go
@@ -11,9 +11,6 @@ import (
 	grpc1 "github.com/gogo/protobuf/grpc"
 	"github.com/gogo/protobuf/proto"
 	"github.com/spf13/cobra"
-	"google.golang.org/grpc"
-
-	"github.com/osmosis-labs/osmosis/v13/osmoutils"
 )
 
 func QueryIndexCmd(moduleName string) *cobra.Command {
@@ -74,37 +71,13 @@ func SimpleQueryCmd[reqP proto.Message, querier any](use string, short string, l
 	return SimpleQueryFromDescriptor[reqP](desc, newQueryClientFn)
 }
 
-type ParamGetter[reqP proto.Message, resP proto.Message] interface {
-	Params(context.Context, reqP, ...grpc.CallOption) (resP, error)
-}
-
-func GetParams[reqP proto.Message, resP proto.Message, querier ParamGetter[reqP, resP]](
-	moduleName string,
+func GetParams[reqP proto.Message, querier any](moduleName string,
 	newQueryClientFn func(grpc1.ClientConn) querier) *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "params [flags]",
-		Short: fmt.Sprintf("Get the params for the x/%s module", moduleName),
-		Args:  cobra.ExactArgs(0),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientQueryContext(cmd)
-			if err != nil {
-				return err
-			}
-			queryClient := newQueryClientFn(clientCtx)
-
-			req := osmoutils.MakeNew[reqP]()
-			res, err := queryClient.Params(cmd.Context(), req)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
+	return SimpleQueryFromDescriptor[reqP](QueryDescriptor{
+		Use:         "params [flags]",
+		Short:       fmt.Sprintf("Get the params for the x/%s module", moduleName),
+		QueryFnName: "Params",
+	}, newQueryClientFn)
 }
 
 func callQueryClientFn[reqP proto.Message, querier any](ctx context.Context, fnName string, req reqP, q querier) (res proto.Message, err error) {
